requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
ua-parser==0.16.1
user-agents==2.2.0
